+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/HolderAgent.kt
@@ -176,22 +176,6 @@ class HolderAgent(
         )
     }
 
-    @Deprecated(
-        "Replace with more general implementation due to increasing number of presentation mechanisms",
-        replaceWith = ReplaceWith("createDefaultPresentationExchangePresentation")
-    )
-    override suspend fun createPresentation(
-        request: PresentationRequestParameters,
-        presentationDefinition: PresentationDefinition,
-        fallbackFormatHolder: FormatHolder?,
-        pathAuthorizationValidator: PathAuthorizationValidator?,
-    ): KmmResult<PresentationResponseParameters.PresentationExchangeParameters> =
-        createDefaultPresentationExchangePresentation(
-            request = request,
-            presentationDefinition = presentationDefinition,
-            fallbackFormatHolder = fallbackFormatHolder,
-        )
-
     private suspend fun createDefaultPresentationExchangePresentation(
         request: PresentationRequestParameters,
         presentationDefinition: PresentationDefinition,
@@ -208,7 +192,6 @@ class HolderAgent(
             )
         )
 
-    @Suppress("DEPRECATION")
     private suspend fun createPresentationExchangePresentation(
         request: PresentationRequestParameters,
         credentialPresentation: CredentialPresentation.PresentationExchangePresentation,
@@ -225,25 +208,8 @@ class HolderAgent(
             throw PresentationException(it)
         }
 
-        createPresentation(
-            request = request,
-            presentationDefinitionId = presentationDefinition.id,
-            presentationSubmissionSelection = presentationCredentialSelection.mapValues {
-                CredentialSubmission(
-                    credential = it.value.credential,
-                    disclosedAttributes = it.value.disclosedAttributes
-                )
-            },
-        ).getOrThrow()
-    }
-
-    @Suppress("OVERRIDE_DEPRECATION") // TODO: make private after removing from interface and change selection value type
-    override suspend fun createPresentation(
-        request: PresentationRequestParameters,
-        presentationDefinitionId: String?,
-        @Suppress("DEPRECATION") presentationSubmissionSelection: Map<String, CredentialSubmission>,
-    ): KmmResult<PresentationResponseParameters.PresentationExchangeParameters> = catching {
-        val submissionList = presentationSubmissionSelection.mapValues {
+        val presentationDefinitionId = presentationDefinition.id
+        val submissionList = presentationCredentialSelection.mapValues {
             PresentationExchangeCredentialDisclosure(
                 credential = it.value.credential,
                 disclosedAttributes = it.value.disclosedAttributes
@@ -270,7 +236,6 @@ class HolderAgent(
             presentationResults = verifiablePresentations,
         )
     }
-
 
     private suspend fun createDCQLPresentation(
         request: PresentationRequestParameters,