+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/AgentTest.kt
@@ -10,6 +10,7 @@ import at.asitplus.openid.dcql.DCQLCredentialQueryInstance
 import at.asitplus.openid.dcql.DCQLCredentialQueryList
 import at.asitplus.openid.dcql.DCQLQuery
 import at.asitplus.wallet.lib.data.ConstantIndex
+import at.asitplus.wallet.lib.data.CredentialPresentation.PresentationExchangePresentation
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest
 import at.asitplus.wallet.lib.data.StatusListToken
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
@@ -73,9 +74,13 @@ class AgentTest : FreeSpec({
     }
 
     "when using presentation exchange" - {
-        val singularPresentationDefinition = PresentationDefinition(
-            id = uuid4().toString(),
-            inputDescriptors = listOf(DifInputDescriptor(id = uuid4().toString()))
+        val singularPresentationDefinition = PresentationExchangePresentation(
+            presentationRequest = CredentialPresentationRequest.PresentationExchangeRequest(
+                PresentationDefinition(
+                    DifInputDescriptor(id = uuid4().toString())
+                ),
+            ),
+            inputDescriptorSubmissions = null,
         )
 
         "simple walk-through success" {
@@ -93,8 +98,9 @@ class AgentTest : FreeSpec({
 
             val presentationParameters = holder.createPresentation(
                 request = PresentationRequestParameters(nonce = challenge, audience = verifierId),
-                presentationDefinition = singularPresentationDefinition,
-            ).getOrThrow()
+                credentialPresentation = singularPresentationDefinition,
+            ).getOrThrow().shouldBeInstanceOf<PresentationResponseParameters.PresentationExchangeParameters>()
+
             val vp = presentationParameters.presentationResults.first()
                 .shouldBeInstanceOf<CreatePresentationResult.Signed>()
             val verified = verifier.verifyPresentationVcJwt(vp.jwsSigned.getOrThrow(), challenge)
@@ -117,8 +123,9 @@ class AgentTest : FreeSpec({
                     nonce = challenge,
                     audience = issuer.keyMaterial.identifier
                 ),
-                presentationDefinition = singularPresentationDefinition,
-            ).getOrThrow()
+                credentialPresentation = singularPresentationDefinition,
+            ).getOrThrow().shouldBeInstanceOf<PresentationResponseParameters.PresentationExchangeParameters>()
+
             val vp = presentationParameters.presentationResults.first()
                 .shouldBeInstanceOf<CreatePresentationResult.Signed>()
             val result = verifier.verifyPresentationVcJwt(vp.jwsSigned.getOrThrow(), challenge)
@@ -178,7 +185,7 @@ class AgentTest : FreeSpec({
                         nonce = challenge,
                         audience = "urn:${uuid4()}"
                     ),
-                    presentationDefinition = singularPresentationDefinition,
+                    credentialPresentation = singularPresentationDefinition,
                 ).getOrNull() shouldBe null
             }
         }
@@ -252,7 +259,7 @@ class AgentTest : FreeSpec({
                     nonce = challenge,
                     audience = "urn:${uuid4()}"
                 ),
-                presentationDefinition = singularPresentationDefinition,
+                credentialPresentation = singularPresentationDefinition,
             ).getOrNull() shouldBe null
         }
 
@@ -267,12 +274,12 @@ class AgentTest : FreeSpec({
             holder.storeCredential(credentials.toStoreCredentialInput()).getOrThrow()
             val presentationParameters = holder.createPresentation(
                 request = PresentationRequestParameters(nonce = challenge, audience = verifierId),
-                presentationDefinition = singularPresentationDefinition,
-            ).getOrNull()
-            presentationParameters.shouldNotBeNull()
+                credentialPresentation = singularPresentationDefinition,
+            ).getOrNull().shouldBeInstanceOf<PresentationResponseParameters.PresentationExchangeParameters>()
+
             val vp = presentationParameters.presentationResults.firstOrNull()
-            vp.shouldNotBeNull()
-            vp.shouldBeInstanceOf<CreatePresentationResult.Signed>()
+                .shouldNotBeNull()
+                .shouldBeInstanceOf<CreatePresentationResult.Signed>()
 
             verifier.verifyPresentationVcJwt(vp.jwsSigned.getOrThrow(), challenge).also {
                 it.shouldBeInstanceOf<Verifier.VerifyPresentationResult.Success>()
@@ -292,12 +299,12 @@ class AgentTest : FreeSpec({
             holder.storeCredential(credentials.toStoreCredentialInput()).getOrThrow()
             val presentationParameters = holder.createPresentation(
                 request = PresentationRequestParameters(nonce = challenge, audience = verifierId),
-                presentationDefinition = singularPresentationDefinition,
-            ).getOrNull()
-            presentationParameters.shouldNotBeNull()
+                credentialPresentation = singularPresentationDefinition,
+            ).getOrNull().shouldBeInstanceOf<PresentationResponseParameters.PresentationExchangeParameters>()
+
             val vp = presentationParameters.presentationResults.firstOrNull()
-            vp.shouldNotBeNull()
-            vp.shouldBeInstanceOf<CreatePresentationResult.Signed>()
+                .shouldNotBeNull()
+                .shouldBeInstanceOf<CreatePresentationResult.Signed>()
 
             val credentialsToRevoke = issuer.issueCredential(
                 DummyCredentialDataProvider.getCredential(